Attempt to expand blocks of paired features in a PairSummaries object.

Attempt to expand blocks of paired features in a PairSummaries object.

BlockExpansion(Pairs,
               GapTolerance = 4L,
               DropSingletons = FALSE,
               Criteria = "PID",
               Floor = 0.5,
               NewPairsOnly = TRUE,
               DBPATH,
               Verbose = FALSE)

Arguments

Pairs

An object of class PairSummaries.

GapTolerance

Integer value indicating the diff between feature IDs that can be tolerated to view features as part of the same block. Set by default to 4L, implying that a single feature missing in a run of pairs will not cause the block to be split. Setting to 3L would imply that a diff of 3 between features, or a gap of 2 features, can be viewed as those features being part of the same block.

DropSingletons

Ignore solo pairs when planning expansion routes. Set to FALSE by default.

Criteria

Either ``PID'' or ``Score'', indicating which metric to use to keep or reject pairs.

Floor

Lower PID limit for keeping a pair that was evaluated during expansion.

NewPairsOnly

Logical indicating whether or not to return only the pairs that were kept from all expansion attempts, or to return a PairSummaries object with the new pairs folded in.

DBPATH

A file or connection pointing to the DECIPHER database supplied to FindSynteny for the original map construction.

Verbose

Logical indicating whether or not to display a progress bar and print the time difference upon completion.

Details

BlockExpansion uses a naive expansion algorithm to attempt to fill in gaps in blocks of paired features and to attempt to expand blocks of paired features.

Value

An object of class PairSummaries.

Author

Nicholas Cooley npc19@pitt.edu

See also

Examples

DBPATH <- system.file("extdata",
                      "Endosymbionts.sqlite",
                      package = "SynExtend")
                      
data("Endosymbionts_Pairs01", package = "SynExtend")
Pairs02 <- BlockExpansion(Pairs = Endosymbionts_Pairs01,
                          NewPairsOnly = FALSE,
                          DBPATH = DBPATH,
                          Verbose = TRUE)
#> ### Genome pair: 1 - 2 ###
#> Index pair 1 - 1
#> Feature set 1 - 533 and 1 - 2047:
#> ================================================================================
#> 
#> ### Genome pair: 1 - 3 ###
#> Index pair 1 - 1
#> Feature set 1 - 533 and 1 - 347:
#> ================================================================================
#> 
#> ### Genome pair: 2 - 3 ###
#> Index pair 1 - 1
#> Feature set 1 - 2047 and 1 - 347:
#> ================================================================================
#> 
#> Time difference of 7.802649 secs