View a Database Table in a Web Browser

Opens an html file in a web browser to show the contents of a table in a database.

Usage

BrowseDB(dbFile,
         htmlFile = paste(tempdir(), "/db.html", sep = ""),
         openURL = interactive(),
         tblName = "Seqs",
         identifier = "",
         limit = -1,
         orderBy = "row_names",
         maxChars = 50,
         clause="")

Arguments

dbFile

A SQLite connection object or a character string specifying the path to the database file.

htmlFile

Character string giving the location where the html file should be written.

openURL

Logical indicating whether the htmlFile should be opened in a web browser.

tblName

Character string specifying the table to view.

identifier

Optional character string used to narrow the search results to those matching a specific identifier. If "" then all identifiers are selected.

limit

Number of results to display. The default (-1) does not limit the number of results.

orderBy

Character string giving the column name for sorting the results. Defaults to the order of entries in the database. Optionally can be followed by " ASC" or " DESC" to specify ascending (the default) or descending order.

maxChars

Maximum number of characters to display in each column.

clause

An optional character string to append to the query as part of a ``where clause''.

Value

Creates an html table containing all the fields of the database table and (if openURL is TRUE) opens it in the web browser for viewing. Returns htmlFile if the html file was written successfully.

References

ES Wright (2016) "Using DECIPHER v2.0 to Analyze Big Biological Sequence Data in R". The R Journal, 8(1), 352-359.

Author

Erik Wright eswright@pitt.edu

Note

If viewing a table containing sequences, the sequences are purposefully not shown in the output.

See also

Examples

db <- system.file("extdata", "Bacteria_175seqs.sqlite", package="DECIPHER")
BrowseDB(db)