Calculate the Distances Between Sequences

Calculates a distance matrix for an XStringSet. Each element of the distance matrix corresponds to the dissimilarity between two sequences in the XStringSet.

DistanceMatrix(myXStringSet,
               type = "matrix",
               includeTerminalGaps = FALSE,
               penalizeGapLetterMatches = TRUE,
               penalizeGapGapMatches = FALSE,
               correction = "none",
               processors = 1,
               verbose = TRUE)

Arguments

myXStringSet

A DNAStringSet, RNAStringSet, or AAStringSet object of aligned sequences.

type

Character string indicating the type of output desired. This should be either "matrix" or "dist". (See value section below.)

includeTerminalGaps

Logical specifying whether or not to include terminal gaps ("-" or "." characters on each end of the sequence) into the calculation of distance.

penalizeGapLetterMatches

Logical specifying whether or not to consider gap-to-letter matches as mismatches. If FALSE, then gap-to-letter matches are not included in the total length used to calculate distance.

penalizeGapGapMatches

Logical specifying whether or not to consider gap-to-gap matches as mismatches. If FALSE (the default), then gap-to-gap matches are not included in the total length used to calculate distance.

correction

The substitution model used for distance correction. This should be (an abbreviation of) either "none", "Jukes-Cantor" (i.e., "JC69"), or "F81". For "F81" letter frequencies are derived from myXStringSet.

processors

The number of processors to use, or NULL to automatically detect and use all available processors.

verbose

Logical indicating whether to display progress.

Details

The uncorrected (correction = "none") distance matrix represents the hamming distance between each of the sequences in myXStringSet. Ambiguity can be represented using the characters of the IUPAC_CODE_MAP for DNAStringSet and RNAStringSet inputs, or using the AMINO_ACID_CODE for an AAStringSet input. For example, the distance between an 'N' and any other nucleotide base is zero. The letters B (N or D), J (I or L), Z (Q or E), and X (any letter) are degenerate in the AMINO_ACID_CODE.

If includeTerminalGaps = FALSE then terminal gaps ("-" or "." characters) are not included in sequence length. This can be faster since only the positions common to each pair of sequences are compared. Sequences with no overlapping region in the alignment are given a value of NA, unless includeTerminalGaps = TRUE, in which case distance is 100%. Masked characters ("+") in either sequence are not considered in distance.

Penalizing gap-to-gap and gap-to-letter mismatches specifies whether to penalize these special mismatch types and include them in the total length when calculating distance. Both "-" and "." characters are interpreted as gaps. The default behavior is to calculate distance as the fraction of positions that differ across the region of the alignment shared by both sequences (not including gap-to-gap matches).

Two correction factors are available, "JC69" and "F81", which are described in MODELS. Both transform raw distance (\(d\)) by $$-E * \log \left( 1 - d / E \right)$$, where $$E = 1 - \sum_{i \in sym} f_i^2$$ and (\(f\)) is the relative frequency of each symbol (\(sym\)). In the "JC69" model symbols are assumed to have equal frequency, whereas in the "F81" model the symbol frequencies are empirically derived from the input myXStringSet. Note that gaps are treated as an additional symbol when penalizeGapLetterMatches or penalizeGapGapMatches is TRUE.

The elements of the distance matrix can be referenced by dimnames corresponding to the names of the XStringSet. Additionally, an attribute named "correction" specifying the method of correction used can be accessed using the function attr.

Value

If type is "matrix", a symmetric matrix where each element is the distance between the sequences referenced by the respective row and column. The dimnames of the matrix correspond to the names of the XStringSet. If type is "dist", an object of class"dist" that contains one triangle of the distance matrix as a vector. Since the distance matrix is symmetric, storing only one triangle is more memory efficient.

Author

Erik Wright eswright@pitt.edu

See also

Examples

# example of using the defaults:
dna <- DNAStringSet(c("ACTG", "ACCG"))
dna
#> DNAStringSet object of length 2:
#>     width seq
#> [1]     4 ACTG
#> [2]     4 ACCG
DistanceMatrix(dna)
#> ================================================================================
#> 
#> Time difference of 0 secs
#> 
#>      [,1] [,2]
#> [1,] 0.00 0.25
#> [2,] 0.25 0.00
#> attr(,"correction")
#> [1] "none"

# changing the output type to "dist":
d <- DistanceMatrix(dna, type="dist")
#> ================================================================================
#> 
#> Time difference of 0 secs
#> 
d
#>      1    2
#> 1 0.00 0.25
#> 2 0.25 0.00
length(d) # minimal memory space required
#> [1] 1
m <- as.matrix(d)
length(m) # more memory space required
#> [1] 4

# supplying an AAStringSet
aa <- AAStringSet(c("ASYK", "ATYK", "CTWN"))
aa
#> AAStringSet object of length 3:
#>     width seq
#> [1]     4 ASYK
#> [2]     4 ATYK
#> [3]     4 CTWN
DistanceMatrix(aa)
#> ================================================================================
#> 
#> Time difference of 0 secs
#> 
#>      [,1] [,2] [,3]
#> [1,] 0.00 0.25 1.00
#> [2,] 0.25 0.00 0.75
#> [3,] 1.00 0.75 0.00
#> attr(,"correction")
#> [1] "none"

# defaults compare intersection of internal ranges:
dna <- DNAStringSet(c("ANGCT-", "-ACCT-"))
dna
#> DNAStringSet object of length 2:
#>     width seq
#> [1]     6 ANGCT-
#> [2]     6 -ACCT-
d <- DistanceMatrix(dna)
#> ================================================================================
#> 
#> Time difference of 0 secs
#> 
d
#>      [,1] [,2]
#> [1,] 0.00 0.25
#> [2,] 0.25 0.00
#> attr(,"correction")
#> [1] "none"
# d[1,2] is 1 base in 4 = 0.25

# compare the entire sequence, including gaps:
dna <- DNAStringSet(c("ANGCT-", "-ACCT-"))
dna
#> DNAStringSet object of length 2:
#>     width seq
#> [1]     6 ANGCT-
#> [2]     6 -ACCT-
d <- DistanceMatrix(dna, includeTerminalGaps=TRUE,
                    penalizeGapGapMatches=TRUE)
#> ================================================================================
#> 
#> Time difference of 0 secs
#> 
d
#>      [,1] [,2]
#> [1,]  0.0  0.5
#> [2,]  0.5  0.0
#> attr(,"correction")
#> [1] "none"
# d[1,2] is now 3 bases in 6 = 0.50

# compare union of internal positions, without terminal gaps:
dna <- DNAStringSet(c("ANGCT-", "-ACCT-"))
dna
#> DNAStringSet object of length 2:
#>     width seq
#> [1]     6 ANGCT-
#> [2]     6 -ACCT-
d <- DistanceMatrix(dna, includeTerminalGaps=TRUE,
                    penalizeGapGapMatches=FALSE)
#> ================================================================================
#> 
#> Time difference of 0 secs
#> 
d
#>      [,1] [,2]
#> [1,]  0.0  0.4
#> [2,]  0.4  0.0
#> attr(,"correction")
#> [1] "none"
# d[1,2] is now 2 bases in 5 = 0.40

# gap ("-") and unknown (".") characters are interchangeable:
dna <- DNAStringSet(c("ANGCT.", ".ACCT-"))
dna
#> DNAStringSet object of length 2:
#>     width seq
#> [1]     6 ANGCT.
#> [2]     6 .ACCT-
d <- DistanceMatrix(dna, includeTerminalGaps=TRUE,
                    penalizeGapGapMatches=FALSE)
#> ================================================================================
#> 
#> Time difference of 0 secs
#> 
d
#>      [,1] [,2]
#> [1,]  0.0  0.4
#> [2,]  0.4  0.0
#> attr(,"correction")
#> [1] "none"
# d[1,2] is still 2 bases in 5 = 0.40