Find Genes in a Genome

Predicts the start and stop positions of protein coding genes in a genome.

Usage

FindGenes(myDNAStringSet,
          geneticCode = getGeneticCode("11"),
          minGeneLength = 60,
          includeGenes = NULL,
          allowEdges = TRUE,
          allScores = FALSE,
          showPlot = FALSE,
          processors = 1,
          verbose = TRUE)

Arguments

myDNAStringSet

A DNAStringSet object of unaligned sequences representing a genome.

geneticCode

A named character vector defining the translation from codons to amino acids. Optionally, an "alt_init_codons" attribute can be used to specify alternative initiation codons. By default, the bacterial and archael genetic code is used, which has seven possible initiation codons: ATG, GTG, TTG, CTG, ATA, ATT, and ATC.

minGeneLength

Integer specifying the minimum length of genes to find in the genome.

includeGenes

A Genes object to include as potential genes or NULL (the default) to predict all genes de novo.

allowEdges

Logical determining whether to allow genes that run off the edge of the sequences. If TRUE (the default), genes can be identified with implied starts or ends outside the boundaries of myDNAStringSet, although the boundary will be set to the last possible codon position. This is useful when genome sequences are circular or incomplete.

allScores

Logical indicating whether to return information about all possible open reading frames or only the predicted genes (the default).

showPlot

Logical determining whether a plot is displayed with the distribution of gene lengths and scores. (See details section below.)

processors

The number of processors to use, or NULL to automatically detect and use all available processors.

verbose

Logical indicating whether to print information about the predictions on each iteration. (See details section below.)

Details

Protein coding genes are identified by learning their characteristic signature directly from the genome, i.e., ab initio prediction. Gene signatures are derived from the content of the open reading frame and surrounding signals that indicate the presence of a gene. Genes are assumed to not contain introns or frame shifts, making the function best suited for prokaryotic genomes.

If showPlot is TRUE then a plot is displayed with four panels. The upper left panel shows the fitted distribution of background open reading frame lengths. The upper right panel shows this distribution on top of the fitted distribution of predicted gene lengths. The lower left panel shows the fitted distribution of scores for the intergenic spacing between genes on the same and opposite genome strands. The bottom right panel shows the total score of open reading frames and predicted genes by length.

If verbose is TRUE, information is shown about the predictions at each iteration of gene finding. The mean score difference between genes and non-genes is updated at each iteration, unless it is negative, in which case the score is dropped and a "-" is displayed. The columns denote the number of iterations ("Iter"), number of codon scoring models ("Models"), start codon scores ("Start"), upstream k-mer motif scores ("Motif"), mRNA folding scores ("Fold"), initial codon bias scores ("Init"), upstream nucleotide bias scores ("UpsNt"), termination codon bias scores ("Term"), ribosome binding site scores ("RBS"), codon autocorrelation scores ("Auto"), stop codon scores ("Stop"), and number of predicted genes ("Genes").

Value

An object of class Genes.

Author

Erik Wright eswright@pitt.edu

See also

Examples

# import a test genome
fas <- system.file("extdata",
  "Chlamydia_trachomatis_NC_000117.fas.gz",
  package="DECIPHER")
genome <- readDNAStringSet(fas)

z <- FindGenes(genome)
#> Iter  Models Start Motif  Init  Fold UpsNt  Term   RBS  Auto  Stop Genes
#> 1          1 18.40                 _                                 8861          1 18.40                 \                                 8861          1 18.40                 |                                 8861          1 18.40                 /                                 8861          1 18.40                 _                                 8861          1 18.40                 \                                 8861          1 18.40                 |                                 8861          1 18.40                 /                                 8861          1 18.40                 _                                 8862         12 18.40  0.48  0.90                                       8883         15 18.38  0.53  1.16  0.40  1.41                           8914         14 15.59  0.52  1.27  0.46  1.76  0.24                     8945         10 15.62  0.52  1.25  0.49  1.84  0.25  0.98  0.10  0.07   8976         11 15.61  0.51  1.24  0.50  1.92  0.25  1.04  0.10  0.07   8977         13 15.59  0.51  1.25  0.51  1.94  0.25  1.05  0.10  0.07   8988         12 15.59  0.51  1.25  0.50  1.94  0.26  1.03  0.10  0.07   8979         12 15.59  0.51  1.25  0.51  1.95  0.26  1.08  0.10  0.07   89710        12 15.58  0.51  1.25  0.49  1.96  0.25  1.03  0.10  0.07   89711        12 15.57  0.51  1.25  0.50  1.96  0.26  1.04  0.10  0.07   89812        13 15.57  0.52  1.25  0.52  1.97  0.26  1.06  0.10  0.07   89713        13 15.57  0.52  1.25  0.51  1.98  0.25  1.08  0.10  0.07   89914        13 15.57  0.52  1.25  0.51  1.98  0.25  1.08  0.10  0.07   897
#> 
#> Time difference of 31.79 secs
#> 
z
#> Genes object of size 897 specifying:
#> 897 protein coding genes from 66 to 5,361 nucleotides.
#> 
#>   Index Strand Begin  End TotalScore ... Gene
#> 1     1      0     1 1176      77.84 ...    1
#> 2     1      1  1321 1593      21.00 ...    1
#> 3     1      0  1794 2096      34.43 ...    1
#> 4     1      0  2108 3583     139.23 ...    1
#> 5     1      0  3585 5051     124.83 ...    1
#> 6     1      1  5150 6241      78.66 ...    1
#> ... with 891 more rows.
genes <- ExtractGenes(z, genome)
genes
#> DNAStringSet object of length 897:
#>       width seq
#>   [1]  1176 GCGGCCGCCCGGGAAATTGCTAAAAGATGGGAG...GGCGTTGGAATAGAGAAGTCGATAGGGAATAA
#>   [2]   273 ATGCTTTGTAAAGTTTGTAGAGGATTATCTTCT...GATACCACCACCATCATATGGATAGAGAATAA
#>   [3]   303 ATGACAGAGTCATATGTAAACAAAGAAGAAATC...GATTAGTCAAAGTCCCTACAGTTATCAAATAG
#>   [4]  1476 ATGTATCGTAAGAGTGCTTTAGAATTAAGAGAT...TGAATGGACTTTTTGACGGAGGAATAGAATAA
#>   [5]  1467 ATGGGCATAGCACATACTGAATGGGAGTCTGTG...AATTGCTATTAGCAGCTATGCGAGATATGTAA
#>   ...   ... ...
#> [893]  3051 ATGCCTTTTTCTTTGAGATCTACATCATTTTGT...TTGTATCCATGGGCTTGAATAGAATCTTTTAA
#> [894]   303 ATGCTGGCAACAATTAAAAAAATTACTGTGTTG...GAGATTCTCGCCTAGAATGCAAGAAGATATAA
#> [895]  2637 ATGCGACCTGATCATATGAACTTCTGTTGTCTA...CCCTGGATCTGGGGACCACTTACAGGTTCTAG
#> [896]    96 ATGTCAAAAAAAAGTAATAATTTACAGACTTTT...ATGAAGAGTTAAGGAAGATTTTTGGATTGTGA
#> [897]   600 ATGAGCATCAGGGGAGTAGGAGGCAACGGGAAT...AAGCGAACAAGTTGGCTGGCCCTGACGGGGTA
proteins <- ExtractGenes(z, genome, type="AAStringSet")
proteins
#> AAStringSet object of length 897:
#>       width seq
#>   [1]   392 AAAREIAKRWEQRVRDLQDKGAARKLLNDPLGR...QVEGILRDMLTNGSQTFRDLMRRWNREVDRE*
#>   [2]    91 MLCKVCRGLSSLIVVLGAINTGILGVTGYKVNL...CLNFLKCCFKKRHGDCCSSKGGYHHHHMDRE*
#>   [3]   101 MTESYVNKEEIISLAKNAALELEDAHVEEFVTS...DMVTSDFTQEEFLSNVPVSLGGLVKVPTVIK*
#>   [4]   492 MYRKSALELRDAVVNRELSVTAITEYFYHRIES...ICQVGYSFQEHSQIKQLYPKAVNGLFDGGIE*
#>   [5]   489 MGIAHTEWESVIGLEVHVELNTESKLFSPARNH...GFLVGQIMKRTEGKAPPKRVNELLLAAMRDM*
#>   ...   ... ...
#> [893]  1017 MPFSLRSTSFCFLACLCSYSYGFASSPQVLTPN...HHFGRAYMNYSLDARRRQTAHFVSMGLNRIF*
#> [894]   101 MLATIKKITVLLLSKRKAGIRIDYCALALDAVE...LDASLESAQVRLAGLMLDYWDGDSRLECKKI*
#> [895]   879 MRPDHMNFCCLCAAILSSTAVLFGQDPLGETAL...LHRLQTLLNVSYVLRGQSHSYSLDLGTTYRF*
#> [896]    32 MSKKSNNLQTFSSRALFHVFQDEELRKIFGL*
#> [897]   200 MSIRGVGGNGNSRIPSHNGDGSNRRSQNTKGNN...NLDVNEARLMAAYTSECADHLEANKLAGPDGV