Find Non-Coding RNAs in a Genome

Searches for conserved patterns representing a family of non-coding RNAs. Returns the start and end boundaries of potential matches along with their log-odds score.

FindNonCoding(x,
              myXStringSet,
              minScore = 16,
              allScores = FALSE,
              processors = 1,
              verbose = TRUE)

Arguments

x

A NonCoding object or a list of NonCoding objects for searching.

myXStringSet

A DNAStringSet or RNAStringSet object of unaligned sequences, typically representing a genome.

minScore

Numeric giving the minimum log-odds score of matches to x in myXStringSet to report, or a vector of numerics specifying the minimum score per NonCoding object in x. The maximum false discovery rate is approximately exp(-minScore) per nucleotide per object in x.

allScores

Logical specifying whether all matches should be returned (TRUE) or only the top matches when there are multiple matches in the same region.

processors

The number of processors to use, or NULL to automatically detect and use all available processors.

verbose

Logical indicating whether to display progress.

Details

Non-coding RNAs are identified by the location of representative sequence patterns relative to the beginning and end of the non-coding RNA. Potential matches to each NonCoding object in x are scored based on their log-odds relative to a background that is derived from the input sequence (myXStringSet). Matches of at least minScore are returned as a Genes object with the "Gene" column set to the negative index of the list element of x that was identified.

Value

An object of class Genes.

References

Wright, E. S. (2021). FindNonCoding: rapid and simple detection of non-coding RNAs in genomes. Bioinformatics. https://doi.org/10.1093/bioinformatics/btab708

Author

Erik Wright eswright@pitt.edu

See also

Examples

data(NonCodingRNA_Bacteria)
x <- NonCodingRNA_Bacteria
names(x)
#>  [1] "tRNA-Ala"                            
#>  [2] "tRNA-Arg"                            
#>  [3] "tRNA-Asn"                            
#>  [4] "tRNA-Asp"                            
#>  [5] "tRNA-Cys"                            
#>  [6] "tRNA-Gln"                            
#>  [7] "tRNA-Glu"                            
#>  [8] "tRNA-Gly"                            
#>  [9] "tRNA-His"                            
#> [10] "tRNA-Ile"                            
#> [11] "tRNA-Leu"                            
#> [12] "tRNA-Lys"                            
#> [13] "tRNA-Met"                            
#> [14] "tRNA-Phe"                            
#> [15] "tRNA-Pro"                            
#> [16] "tRNA-Ser"                            
#> [17] "tRNA-Thr"                            
#> [18] "tRNA-Trp"                            
#> [19] "tRNA-Tyr"                            
#> [20] "tRNA-Val"                            
#> [21] "tRNA-Sec"                            
#> [22] "rRNA_5S-RF00001"                     
#> [23] "rRNA_16S-RF00177"                    
#> [24] "rRNA_23S-RF02541"                    
#> [25] "tmRNA-RF00023"                       
#> [26] "tmRNA_Alpha-RF01849"                 
#> [27] "RNase_P_class_A-RF00010"             
#> [28] "RNase_P_class_B-RF00011"             
#> [29] "SsrS-RF00013"                        
#> [30] "Intron_Gp_I-RF00028"                 
#> [31] "Intron_Gp_II-RF00029"                
#> [32] "SmallSRP-RF00169"                    
#> [33] "Cyclic-di-GMP_Riboswitch-RF01051"    
#> [34] "Cyclic-di-AMP_Riboswitch-RF00379"    
#> [35] "T-box_Leader-RF00230"                
#> [36] "Ribosomal_Protein_L10_Leader-RF00557"
#> [37] "Cobalamin_Riboswitch-RF00174"        
#> [38] "TPP_Riboswitch-RF00059"              
#> [39] "SAM_Riboswitch-RF00162"              
#> [40] "Fluoride_Riboswitch-RF01734"         
#> [41] "FMN_Riboswitch-RF00050"              
#> [42] "Glycine_Riboswitch-RF00504"          
#> [43] "HEARO-RF02033"                       
#> [44] "Flavo_1-RF01705"                     
#> [45] "Acido_Lenti_1-RF01687"               
#> [46] "5'_ureB-RF02514"                     

# import a test genome
fas <- system.file("extdata",
  "Chlamydia_trachomatis_NC_000117.fas.gz",
  package="DECIPHER")
genome <- readDNAStringSet(fas)

z <- FindNonCoding(x, genome)
#> ================================================================================
#> 
#> Time difference of 50.76 secs
z
#> Genes object of size 46 specifying:
#> 46 non-coding RNAs from 72 to 2,938 nucleotides.
#> 
#>   Index Strand  Begin    End TotalScore Gene
#> 1     1      1  20663  21082      88.33  -25
#> 2     1      1  42727  42801      69.94   -3
#> 3     1      1  68920  68995      78.62  -15
#> 4     1      0 158662 158736      55.75  -17
#> 5     1      0 158744 158827      52.31  -19
#> 6     1      1 202339 202414      68.64  -10
#> ... with 40 more rows.

annotations <- attr(z, "annotations")
m <- match(z[, "Gene"], annotations)
sort(table(names(annotations)[m]))
#> 
#> RNase_P_class_A-RF00010        SmallSRP-RF00169                tRNA-Asn 
#>                       1                       1                       1 
#>                tRNA-Asp                tRNA-Cys                tRNA-Gln 
#>                       1                       1                       1 
#>                tRNA-Glu                tRNA-His                tRNA-Ile 
#>                       1                       1                       1 
#>                tRNA-Lys                tRNA-Phe                tRNA-Trp 
#>                       1                       1                       1 
#>                tRNA-Tyr           tmRNA-RF00023        rRNA_16S-RF00177 
#>                       1                       1                       2 
#>        rRNA_23S-RF02541         rRNA_5S-RF00001                tRNA-Ala 
#>                       2                       2                       2 
#>                tRNA-Gly                tRNA-Pro                tRNA-Val 
#>                       2                       2                       2 
#>                tRNA-Arg                tRNA-Met                tRNA-Thr 
#>                       3                       3                       3 
#>                tRNA-Ser                tRNA-Leu 
#>                       4                       5 

genes <- ExtractGenes(z, genome, type="RNAStringSet")
genes
#> RNAStringSet object of length 46:
#>      width seq
#>  [1]   420 GGGGGUGUAAAGGUUUCGACUUAGAAAUGAAGC...AGGACGAGAGUUCGACUCUCUCCACCUCCAUAG
#>  [2]    75 UCCGGAGUAGCUCAGCGGUAGAGCAGUGGACUG...UGGUCGUUGGUUCGAACCCAUCCUCCGGAGUCU
#>  [3]    76 CGGAGUAUAGCGCAGCCUGGUUAGCGCGGUUGC...AUAGGUCGGGGGUUCGAAUCCCUCUACUCCGAU
#>  [4]    75 GCUGGAGUAGCUCAAUUGGCAGAGCAUUCGAUU...ACGGUUGAGGGUUCAAUUCCUUUCUCCAGCAUC
#>  [5]    84 GGGGGUGUCGCAUAGCGGUCAAUUGCAUCGGAC...CGGAUACGUUGGUUCAAAUCCAGCCACCCCCAG
#>  ...   ... ...
#> [42]    72 GGUGGCAUCGCCAAGCGGUAAGGCCGAGGCCUG...CUCUAUCCCCGGUUCGAUUCCGGGUGCCACCUU
#> [43]    74 UGGGGUGUGGCCAAGCGGUAAGGCAGCGGUUUU...CGCAUCGGAGGUUCGAAUCCUUCCACCCCAGAG
#> [44]    75 GGGGUAUUAGCUCAGUUGGUUAGAGCGUCACGU...GAAGGUCAGCUGUUCAAGUCAGCUAUAUCCCAA
#> [45]    88 GGAAGAAUGGCAGAGCGGUUUAAUGCACCUGUC...GGUCCGGGGGUUCGAAUCCCUCUUCUUCCGCAU
#> [46]    85 GCCCAGGUGGUGAAAUUGGUAGACACGCUGGAU...GGCAUGUAGGUUCAAGUCCUAUCCUGGGCAUAG