Cluster Sequences By Similarity

Groups the sequences into approximate clusters of similarity.

Usage

IdClusters(myXStringSet,
           cutoff = 0,
           processors = 1,
           verbose = TRUE)

Arguments

myXStringSet

If method is "ML" or reconstruct is not FALSE, the DNAStringSet or RNAStringSet used in the creation of myDistMatrix. If method is "inexact", the DNAStringSet, RNAStringSet, or AAStringSet to cluster.

cutoff

A vector with the maximum edge length separating the sequences in the same cluster. Multiple cutoffs may be provided in ascending or descending order. (See details section below.)

processors

The number of processors to use, or NULL to automatically detect and use all available processors.

verbose

Logical indicating whether to display progress.

Details

IdClusters groups the input sequences into approximate clusters using a heuristic algorithm. First the sequences are ordered by length and the longest sequence becomes the first cluster seed. If the second sequence is less than cutoff k-mer distance then it is added to the cluster, otherwise it becomes a new cluster representative. The remaining sequences are matched to cluster representatives in a similar fashion until all sequences belong to a cluster. In the majority of cases, this process results in clusters with members separated by less than cutoff distance.

Multiple cutoffs may be provided in sorted order. If the cutoffs are provided in descending order then clustering at each new value of cutoff is continued within the prior cutoff's clusters. In this way clusters at lower values of cutoff are completely contained within their umbrella clusters at higher values of cutoff. This is useful for accelerating the clustering process, because each subsequent group is only clustered within the previous group. If multiple cutoffs are provided in ascending order then clustering at each level of cutoff is independent of the prior level.

Value

A data.frame is returned with dimensions \(N*M\), where each one of \(N\) sequences is assigned to a cluster at the \(M\)-level of cutoff. The row.names of the data.frame correspond to the names of myXStingSet.

Author

Erik Wright eswright@pitt.edu

Examples

fas <- system.file("extdata", "50S_ribosomal_protein_L2.fas", package="DECIPHER")
dna <- readDNAStringSet(fas)
aa <- translate(dna)
inexact <- IdClusters(aa, cutoff=seq(0.7, 0.1, -0.1))
#> ================================================================================
#> Time difference of 0.07 secs
#> 
head(inexact)
#>                                     cluster_7 cluster_6 cluster_5 cluster_4
#> Rickettsia prowazekii str. Dachau           8        11        23        21
#> Porphyromonas gingivalis W83               12         5         3         6
#> Porphyromonas gingivalis TDC60             12         5         3         6
#> Porphyromonas gingivalis ATCC 33277        12         5         3         6
#> Pasteurella multocida 671/90               13         4         0         2
#> Pasteurella multocida 36950                13         4         0         2
#>                                     cluster_3 cluster_2 cluster_1
#> Rickettsia prowazekii str. Dachau          27        24        30
#> Porphyromonas gingivalis W83                3         7         3
#> Porphyromonas gingivalis TDC60              3         7         3
#> Porphyromonas gingivalis ATCC 33277         3         7         3
#> Pasteurella multocida 671/90               15         2        94
#> Pasteurella multocida 36950                15         2        94
apply(inexact, 2, max) # number of clusters per cutoff
#> cluster_7 cluster_6 cluster_5 cluster_4 cluster_3 cluster_2 cluster_1 
#>        15        33        53        71        85        94        98