Learn a Non-Coding RNA Model

Learns a compact representation of patterns representing a set of non-coding RNAs belonging to the same family.

LearnNonCoding(myXStringSet,
               threshold = 0.3,
               weight = NA,
               maxLoopLength = 500,
               maxPatterns = 20,
               scoreDependence = FALSE,
               structure = NULL,
               processors = 1)

Arguments

myXStringSet

A DNAStringSet or RNAStringSet object of aligned sequence representatives belonging to the same non-coding RNA family.

threshold

Numeric specifying the minimum relative frequency of patterns to consider during learning.

weight

Either a numeric vector of weights for each sequence, a single number implying equal weights, or NA (the default) to automatically calculate sequence weights based on myXStringSet.

maxLoopLength

Numeric giving the maximum length of conserved hairpin loops to consider.

maxPatterns

A numeric vector of length two specifying the maximum number of motifs and hairpins, respectively, or a single numeric giving the maximum for each.

scoreDependence

Logical determining whether to record a log-odds score for dependencies between patterns. The default (FALSE) is recommended for most non-coding RNA families.

structure

Either a character string providing the consensus secondary structure in dot bracket notation, a matrix of paired positions in the first two columns, or NULL (the default) to predict the consensus secondary structure with PredictDBN.

processors

The number of processors to use, or NULL to automatically detect and use all available processors.

Details

Non-coding RNAs belonging to the same family typically have conserved sequence motifs, secondary structure elements, and k-mer frequencies that can be used to identify members of the family. LearnNonCoding identifies these conserved patterns and determines which are best for identifying the non-coding RNA relative to a random sequence background. Sequence motifs and hairpins are defined relative to their distance from the start or end of the non-coding RNA, allowing the precise and rapid identification of the boundaries of any matches to the non-coding RNA in a genome.

Value

An object of class NonCoding.

References

Wright, E. S. (2021). FindNonCoding: rapid and simple detection of non-coding RNAs in genomes. Bioinformatics. https://doi.org/10.1093/bioinformatics/btab708

Author

Erik Wright eswright@pitt.edu

See also

Examples

# import a family of non-coding RNAs
fas_path <- system.file("extdata",
  "IhtA.fas",
  package="DECIPHER")
rna <- readRNAStringSet(fas_path)
rna
#> RNAStringSet object of length 27:
#>      width seq                                              names               
#>  [1]   103 AAGAUGAUAUUCUGCGCCAUGGA...UGAUUGUUGUUUUCUUGGCUUU KE360988.1/6443-6...
#>  [2]   105 AAGAUGAUAUUCUCCGCCGUGGA...AACUUUAUGUUUUCUUGGCUUU AE002161.1/54676-...
#>  [3]   107 AAGUUGGUAUUCUAACGCCAUGG...UAUCUCCGGUUCUCUUGGCUUU AE001273.1/773281...
#>  [4]   107 AAGUUGGUAUUCUAACGCCAUGG...UGUCUCCAGUUCUCUUGGCUUU AE002160.2/53054-...
#>  [5]   103 AAGAUGAUAUUCUACGCCAUGGA...UGAUUGUUGUUUUCUUGGCUUU AE015925.1/52097-...
#>  ...   ... ...
#> [23]   107 AAGUUGGUAUUCUAACGCCAUGG...UAUCUCCGGUUCUCUUGGCUUU CP006674.1 Chlamy...
#> [24]   103 AAGAUGAUAUUCUGCGCCAUGGA...UGAUUGUUGUUUUCUUGGCUUU KE356008.1 Chlamy...
#> [25]   108 AAGUUGGUAUUCUAACGCCAUGG...UAUCUCCGGUUCUCUUGGCUUU CVNC01000001.1 Ch...
#> [26]   103 AAGAUGAUAUUCUGCGCCAUGGA...UGAUUGUUGUUUUCUUGGCUUU AP006861.1 Chlamy...
#> [27]   101 AAGAUGAUAUUCUACGCCAUGGA...AUGAUGUUGUUUUCUUGGCUUU CP015840.1 Chlamy...

# align the sequences
RNA <- AlignSeqs(rna)
#> Determining distance matrix based on shared 7-mers:
#> ================================================================================
#> 
#> Time difference of 0 secs
#> 
#> Clustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0 secs
#> 
#> Aligning Sequences:
#> ================================================================================
#> 
#> Time difference of 0.13 secs
#> 
#> Iteration 1 of 2:
#> 
#> Determining distance matrix based on alignment:
#> ================================================================================
#> 
#> Time difference of 0 secs
#> 
#> Reclustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.01 secs
#> 
#> Realigning Sequences:
#> ================================================================================
#> 
#> Time difference of 0.05 secs
#> 
#> Iteration 2 of 2:
#> 
#> Determining distance matrix based on alignment:
#> ================================================================================
#> 
#> Time difference of 0 secs
#> 
#> Reclustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.01 secs
#> 
#> Realigning Sequences:
#> ================================================================================
#> 
#> Time difference of 0.01 secs
#> 
RNA
#> RNAStringSet object of length 27:
#>      width seq                                              names               
#>  [1]   113 AAGAUGAUAUUCU-GCGCCAUGG...GAUUGU-UGUUUUCUUGGCUUU KE360988.1/6443-6...
#>  [2]   113 AAGAUGAUAUUCU-CCGCCGUGG...AACUUUAUGUUUUCUUGGCUUU AE002161.1/54676-...
#>  [3]   113 AAGUUGGUAUUCUAACGCCAUGG...--UCUCCGGUUCUCUUGGCUUU AE001273.1/773281...
#>  [4]   113 AAGUUGGUAUUCUAACGCCAUGG...--UCUCCAGUUCUCUUGGCUUU AE002160.2/53054-...
#>  [5]   113 AAGAUGAUAUUCU-ACGCCAUGG...GAUUGU-UGUUUUCUUGGCUUU AE015925.1/52097-...
#>  ...   ... ...
#> [23]   113 AAGUUGGUAUUCUAACGCCAUGG...--UCUCCGGUUCUCUUGGCUUU CP006674.1 Chlamy...
#> [24]   113 AAGAUGAUAUUCU-GCGCCAUGG...GAUUGU-UGUUUUCUUGGCUUU KE356008.1 Chlamy...
#> [25]   113 AAGUUGGUAUUCUAACGCCAUGG...--UCUCCGGUUCUCUUGGCUUU CVNC01000001.1 Ch...
#> [26]   113 AAGAUGAUAUUCU-GCGCCAUGG...GAUUGU-UGUUUUCUUGGCUUU AP006861.1 Chlamy...
#> [27]   113 AAGAUGAUAUUCU-ACGCCAUGG...GA-UGU-UGUUUUCUUGGCUUU CP015840.1 Chlamy...

y <- LearnNonCoding(RNA)
y
#> NonCoding object with 16 motifs, 3 hairpins, and 2-mer frequencies.
y[["motifs"]]
#>    begin_low begin_high end_low end_high         motif          pwm
#> 1          0          0      90       94 AAGaTGrTATTCT 0.945645....
#> 2         13         14      82       86      ACGCCATG 0.796028....
#> 3         21         22      81       85             G 0.018118....
#> 4         22         23      72       76     AATAGCTTC 0.911404....
#> 5         31         33      59       63  rACTCTGkTGyw 0.463332....
#> 6         44         46      56       60            TT 0.042292....
#> 7         46         51      54       58            CA 0.083772....
#> 8         48         53      48       52        GGGGGA 0.018118....
#> 9         54         59      39       43     AAGCCAAGA 0.945645....
#> 10        64         70      34       38           rwr 0.589327....
#> 11        68         74      20       24 rdmGTydGCCGTA 0.554448....
#> 12        82         86      19       20             A 0.945645....
#> 13        84         87      18       19             y 0.018452....
#> 14        85         88      15       16           Tkt 0.018452....
#> 15        88         92       8        8       tGTTTTC 0.164952....
#> 16        95         99       0        0      TTGGCTTT 0.018118....
#>        minscore   prevalence   background
#> 1  0, 10.50.... 0.035714.... 0.996318....
#> 2  0, 4.732.... 0.035714.... 0.987988....
#> 3        0, Inf 0.053671.... 0.550874....
#> 4  0, 7.963.... 0.053671.... 0.981401....
#> 5  0, 6.745.... 0.035714.... 0.975706....
#> 6  0.731723.... 0.120251.... 0.708951....
#> 7        0, Inf 0.064669.... 0.462910....
#> 8  0, 6.420.... 0.064669.... 0.902690....
#> 9  0, 9.116.... 0.064669.... 0.977723....
#> 10 1.471911.... 0.396535.... 0.689376....
#> 11 0, 7.172.... 0.035714.... 0.972311....
#> 12       0, Inf 0.110082.... 0.580776....
#> 13 0.486998.... 0.064669.... 0.449668....
#> 14 0, 0.560.... 0.035714.... 0.641917....
#> 15 0, 4.914.... 0.035714.... 0.948266....
#> 16       0, Inf 0.035714.... 0.990604....
y[["hairpins"]]
#>   begin_low begin_high end_low end_high width_low width_high length_low
#> 1        -4          1      63       72        31         44          6
#> 2        30         37      24       27        43         51          6
#> 3        49         58      -4        0        49         61         13
#>   length_high           dG   prevalence   background
#> 1          12 -Inf, -2.... 0.308347.... 0.210332....
#> 2          15 -Inf, -1.... 0.067775.... 0.000315....
#> 3          26 -Inf, -2.... 0.137634.... 0.000105....
head(y[["kmers"]])
#> [1] 186 109 182 163  78  83