Extract information from a ProtWeb object

ProtWeb objects are outputted from predict.ProtWeaver.

This function extracts the underlying data from the object.

# S3 method for ProtWeb
GetProtWebData(x, AsDf=FALSE, ...)

Arguments

x

A ProtWeb object

AsDf

Should data be printed as a pairwise entry? If TRUE, returns a matrix with three columns, where the first two columns define the pair of genes/proteins and the third column defines the prediction/score. If FALSE, returns an adjacency matrix encoding the same information (but may be sparse depending on how many predictions were made in the original predict call.)

...

Additional parameters for consistency with generic.

Details

predict.ProtWeaver returns a ProtWeb object, which bundles some methods to make formatting and printing of results slightly nicer. This method extracts data form the ProtWeb object.

If AsDf=TRUE, the return data is a 3xN data.frame, with columns 3 showing the prediction for the pair of genes/proteins specified in columns 1 and 2. This format is when predictions are made on small number of pairs (meaning the resulting adjacency matrix is sparse).

If AsDf=FALSE, the return data is a NxN adjacency matrix, with entry i,j containing the prediction for genes i and j.

Value

Either a data.frame or a matrix.

Author

Aidan Lakshman ahl27@pitt.edu

See also

Examples

##############
## Prediction with built-in model and data 
###############

exData <- get(data("ExampleStreptomycesData"))

# Subset isn't necessary but is faster for a working example
pw <- ProtWeaver(exData$Genes[1:10])
#> Warning: Disabling MirrorTree-based algorithms. Input list must include inputs of type "dendrogram" for MT algorithms. Consult the documentation for more info.

protweb <- predict(pw, method='Jaccard')
#> Calculating P/A profiles:
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> Running ProfileDCA:
#>   Finding topology...
#>   || Done.
#>   Eliminating edges close to zero (may take a moment)...
#>   Refining edge weights...
#>   || Done.
#> Running Jaccard:
#>   |                                                                              |                                                                      |   0%  |                                                                              |===                                                                   |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  13%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  22%  |                                                                              |=================                                                     |  24%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  33%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  38%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  42%  |                                                                              |===============================                                       |  44%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  49%  |                                                                              |====================================                                  |  51%  |                                                                              |=====================================                                 |  53%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  58%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  69%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  73%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  78%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%
#> Running Hamming:
#>   |                                                                              |                                                                      |   0%  |                                                                              |===                                                                   |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  13%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  22%  |                                                                              |=================                                                     |  24%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  33%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  38%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  42%  |                                                                              |===============================                                       |  44%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  49%  |                                                                              |====================================                                  |  51%  |                                                                              |=====================================                                 |  53%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  58%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  69%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  73%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  78%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%
#> Running MutualInformation:
#>   |                                                                              |                                                                      |   0%  |                                                                              |===                                                                   |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  13%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  22%  |                                                                              |=================                                                     |  24%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  33%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  38%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  42%  |                                                                              |===============================                                       |  44%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  49%  |                                                                              |====================================                                  |  51%  |                                                                              |=====================================                                 |  53%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  58%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  69%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  73%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  78%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%
#> Running Coloc:
#>   |                                                                              |                                                                      |   0%  |                                                                              |==                                                                    |   2%  |                                                                              |===                                                                   |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  13%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  22%  |                                                                              |=================                                                     |  24%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  33%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  38%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  42%  |                                                                              |===============================                                       |  44%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  49%  |                                                                              |====================================                                  |  51%  |                                                                              |=====================================                                 |  53%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  58%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  69%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  73%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  78%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%
#> Calculating additional P/A Statistics...
#> Predicting with Ensemble method...
#> Done.
#> 
#> Time difference of 2.08 seconds.

# print out results as an adjacency matrix
GetProtWebData(protweb)
#>            1          2          3          4          5         6          7
#> 1  1.0000000 0.39977312 0.40539180 0.28737689 0.15859257 0.5972944 0.16808461
#> 2  0.3997731 1.00000000 0.13100765 0.15225417 0.09975717 0.2330527 0.08474520
#> 3  0.4053918 0.13100765 1.00000000 0.99982883 0.07591308 0.2454367 0.07923918
#> 4  0.2873769 0.15225417 0.99982883 1.00000000 0.06367025 0.1359477 0.10070421
#> 5  0.1585926 0.09975717 0.07591308 0.06367025 1.00000000 0.1118726 0.06912215
#> 6  0.5972944 0.23305267 0.24543674 0.13594773 0.11187258 1.0000000 0.13920004
#> 7  0.1680846 0.08474520 0.07923918 0.10070421 0.06912215 0.1392000 1.00000000
#> 8  0.1069398 0.11237956 0.13187903 0.15281702 0.06119419 0.1253973 0.99748139
#> 9  0.1094592 0.08505273 0.07844418 0.06812339 0.06591927 0.1040911 0.94517017
#> 10 0.1416024 0.08093309 0.07717580 0.06888473 0.11714603 0.0949550 0.76167646
#>             8          9         10
#> 1  0.10693984 0.10945921 0.14160237
#> 2  0.11237956 0.08505273 0.08093309
#> 3  0.13187903 0.07844418 0.07717580
#> 4  0.15281702 0.06812339 0.06888473
#> 5  0.06119419 0.06591927 0.11714603
#> 6  0.12539726 0.10409108 0.09495500
#> 7  0.99748139 0.94517017 0.76167646
#> 8  1.00000000 0.90661135 0.63626273
#> 9  0.90661135 1.00000000 0.99963920
#> 10 0.63626273 0.99963920 1.00000000

# print out results as a pairwise data.frame
GetProtWebData(protweb, AsDf=TRUE)
#>    Gene1 Gene2 Prediction
#> 1      1     2 0.39977312
#> 2      1     3 0.40539180
#> 3      2     3 0.13100765
#> 4      1     4 0.28737689
#> 5      2     4 0.15225417
#> 6      3     4 0.99982883
#> 7      1     5 0.15859257
#> 8      2     5 0.09975717
#> 9      3     5 0.07591308
#> 10     4     5 0.06367025
#> 11     1     6 0.59729438
#> 12     2     6 0.23305267
#> 13     3     6 0.24543674
#> 14     4     6 0.13594773
#> 15     5     6 0.11187258
#> 16     1     7 0.16808461
#> 17     2     7 0.08474520
#> 18     3     7 0.07923918
#> 19     4     7 0.10070421
#> 20     5     7 0.06912215
#> 21     6     7 0.13920004
#> 22     1     8 0.10693984
#> 23     2     8 0.11237956
#> 24     3     8 0.13187903
#> 25     4     8 0.15281702
#> 26     5     8 0.06119419
#> 27     6     8 0.12539726
#> 28     7     8 0.99748139
#> 29     1     9 0.10945921
#> 30     2     9 0.08505273
#> 31     3     9 0.07844418
#> 32     4     9 0.06812339
#> 33     5     9 0.06591927
#> 34     6     9 0.10409108
#> 35     7     9 0.94517017
#> 36     8     9 0.90661135
#> 37     1    10 0.14160237
#> 38     2    10 0.08093309
#> 39     3    10 0.07717580
#> 40     4    10 0.06888473
#> 41     5    10 0.11714603
#> 42     6    10 0.09495500
#> 43     7    10 0.76167646
#> 44     8    10 0.63626273
#> 45     9    10 0.99963920