Align Sequences By Their Amino Acid Translation

Performs alignment of a set of DNA or RNA sequences by aligning their corresponding amino acid sequences.

Usage

AlignTranslation(myXStringSet,
                 sense = "+",
                 direction = "5' to 3'",
                 readingFrame = NA,
                 type = class(myXStringSet),
                 geneticCode = GENETIC_CODE,
                 ...)

Arguments

myXStringSet

A DNAStringSet or RNAStringSet object of unaligned sequences.

sense

Single character specifying sense of the input sequences, either the positive ("+") coding strand or negative ("-") non-coding strand.

direction

Direction of the input sequences, either "5' to 3'" or "3' to 5'".

readingFrame

Numeric vector giving a single reading frame for all of the sequences, or an individual reading frame for each sequence in myXStringSet. The readingFrame can be either 1, 2, 3 to begin translating on the first, second, and third nucleotide position, or NA (the default) to guess the reading frame. (See details section below.)

type

Character string indicating the type of output desired. This should be (an abbreviation of) one of "DNAStringSet", "RNAStringSet", "AAStringSet", or "both". (See value section below.)

geneticCode

Either a character vector giving the genetic code in the same format as GENETIC_CODE (the default), or a list containing one genetic code for each sequence in myXStringSet.

...

Further arguments to be passed directly to AlignSeqs, including gapOpening, gapExtension, gapPower, terminalGap, restrict, anchor, normPower, substitutionMatrix, structureMatrix, standardize, alphabet, guideTree, iterations, refinements, useStructures, structures, FUN, and levels.

Details

Alignment of proteins is often more accurate than alignment of their coding nucleic acid sequences. This function aligns the input nucleic acid sequences via aligning their translated amino acid sequences. First, the input sequences are translated according to the specified sense, direction, and readingFrame. The resulting amino acid sequences are aligned using AlignSeqs, and this alignment is (conceptually) reverse translated into the original sequence type, sense, and direction. Not only is alignment of protein sequences generally more accurate, but aligning translations will ensure that the reading frame is maintained in the nucleotide sequences.

If the readingFrame is NA (the default) then an attempt is made to guess the reading frame of each sequence based on the number of stop codons in the translated amino acids. For each sequence, the first reading frame will be chosen (either 1, 2, or 3) without stop codons, except in the final position. If the number of stop codons is inconclusive for a sequence then the reading frame will default to 1. The entire length of each sequence is translated in spite of any stop codons identified. Note that this method is only constructive in circumstances where there is a substantially long coding sequence with at most a single stop codon expected in the final position, and therefore it is preferable to specify the reading frame of each sequence if it is known.

Value

An XStringSet of the class specified by type, or a list of two components (nucleotides and amino acids) if type is "both". Note that incomplete starting and ending codons will be translated into the mask character ("+") if the result includes an AAStringSet.

References

Wright, E. S. (2015). DECIPHER: harnessing local sequence context to improve protein multiple sequence alignment. BMC Bioinformatics, 16, 322. http://doi.org/10.1186/s12859-015-0749-z

Author

Erik Wright eswright@pitt.edu

See also

Examples

# first three sequences translate to  MFITP*
# and the last sequence translates as MF-TP*
rna <- RNAStringSet(c("AUGUUCAUCACCCCCUAA", "AUGUUCAUAACUCCUUGA",
  "AUGUUCAUUACACCGUAG", "AUGUUUACCCCAUAA"))
RNA <- AlignSeqs(rna, verbose=FALSE)
RNA
#> RNAStringSet object of length 4:
#>     width seq
#> [1]    18 AUGUUCAUCACCCCCUAA
#> [2]    18 AUGUUCAUAACUCCUUGA
#> [3]    18 AUGUUCAUUACACCGUAG
#> [4]    18 AUGU---UUACCCCAUAA

RNA <- AlignTranslation(rna, verbose=FALSE)
RNA
#> RNAStringSet object of length 4:
#>     width seq
#> [1]    18 AUGUUCAUCACCCCCUAA
#> [2]    18 AUGUUCAUAACUCCUUGA
#> [3]    18 AUGUUCAUUACACCGUAG
#> [4]    18 ---AUGUUUACCCCAUAA

AA <- AlignTranslation(rna, type="AAStringSet", verbose=FALSE)
AA
#> AAStringSet object of length 4:
#>     width seq
#> [1]     6 MFITP*
#> [2]     6 MFITP*
#> [3]     6 MFITP*
#> [4]     6 -MFTP*

BOTH <- AlignTranslation(rna, type="both", verbose=FALSE)
BOTH
#> [[1]]
#> RNAStringSet object of length 4:
#>     width seq
#> [1]    18 AUGUUCAUCACCCCCUAA
#> [2]    18 AUGUUCAUAACUCCUUGA
#> [3]    18 AUGUUCAUUACACCGUAG
#> [4]    18 ---AUGUUUACCCCAUAA
#> 
#> [[2]]
#> AAStringSet object of length 4:
#>     width seq
#> [1]     6 MFITP*
#> [2]     6 MFITP*
#> [3]     6 MFITP*
#> [4]     6 -MFTP*
#> 

# example of aligning many protein coding sequences:
fas <- system.file("extdata", "50S_ribosomal_protein_L2.fas", package="DECIPHER")
dna <- readDNAStringSet(fas)
DNA <- AlignTranslation(dna) # align the translation then reverse translate
#> Determining distance matrix based on shared 5-mers:
#> ================================================================================
#> 
#> Time difference of 0.55 secs
#> 
#> Clustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.05 secs
#> 
#> Aligning Sequences:
#> ================================================================================
#> 
#> Time difference of 1.48 secs
#> 
#> Iteration 1 of 2:
#> 
#> Determining distance matrix based on alignment:
#> ================================================================================
#> 
#> Time difference of 0.09 secs
#> 
#> Reclustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.07 secs
#> 
#> Realigning Sequences:
#> ================================================================================
#> 
#> Time difference of 0.65 secs
#> 
#> Iteration 2 of 2:
#> 
#> Determining distance matrix based on alignment:
#> ================================================================================
#> 
#> Time difference of 0.09 secs
#> 
#> Reclustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.08 secs
#> 
#> Realigning Sequences:
#> ================================================================================
#> 
#> Time difference of 0.09 secs
#> 
DNA
#> DNAStringSet object of length 317:
#>       width seq                                             names               
#>   [1]  1011 ----------------------...---------------------- Rickettsia prowaz...
#>   [2]  1011 ----------------------...---------------------- Porphyromonas gin...
#>   [3]  1011 ----------------------...---------------------- Porphyromonas gin...
#>   [4]  1011 ----------------------...---------------------- Porphyromonas gin...
#>   [5]  1011 ----------------------...---------------------- Pasteurella multo...
#>   ...   ... ...
#> [313]  1011 ----------------------...---------------------- Pectobacterium at...
#> [314]  1011 ----------------------...---------------------- Acinetobacter sp....
#> [315]  1011 ----------------------...GGGTCGCGGTGGTCGTCAGTCT Thermosynechococc...
#> [316]  1011 ----------------------...---------------------- Bradyrhizobium ja...
#> [317]  1011 ----------------------...AGGT------------------ Gloeobacter viola...

# using a mixture of standard and non-standard genetic codes
gC1 <- getGeneticCode(id_or_name2="1", full.search=FALSE, as.data.frame=FALSE)
# Mollicutes use an alternative genetic code
gC2 <- getGeneticCode(id_or_name2="4", full.search=FALSE, as.data.frame=FALSE)
w <- grep("Mycoplasma|Ureaplasma", names(dna))
gC <- vector("list", length(dna))
gC[-w] <- list(gC1)
gC[w] <- list(gC2)
AA <- AlignTranslation(dna, geneticCode=gC, type="AAStringSet")
#> Determining distance matrix based on shared 5-mers:
#> ================================================================================
#> 
#> Time difference of 0.56 secs
#> 
#> Clustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.05 secs
#> 
#> Aligning Sequences:
#> ================================================================================
#> 
#> Time difference of 1.47 secs
#> 
#> Iteration 1 of 2:
#> 
#> Determining distance matrix based on alignment:
#> ================================================================================
#> 
#> Time difference of 0.09 secs
#> 
#> Reclustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.07 secs
#> 
#> Realigning Sequences:
#> ================================================================================
#> 
#> Time difference of 0.65 secs
#> 
#> Iteration 2 of 2:
#> 
#> Determining distance matrix based on alignment:
#> ================================================================================
#> 
#> Time difference of 0.09 secs
#> 
#> Reclustering into groups by similarity:
#> ================================================================================
#> 
#> Time difference of 0.1 secs
#> 
#> Realigning Sequences:
#> ================================================================================
#> 
#> Time difference of 0.09 secs
#> 
BrowseSeqs(AA)