Extract information from a ProtWeb object

ProtWeb objects are outputted from predict.ProtWeaver.

This function extracts the underlying data from the object.

Usage

# S3 method for ProtWeb
GetProtWebData(x, AsDf=FALSE, ...)

Arguments

x

A ProtWeb object

AsDf

Should data be printed as a pairwise entry? If TRUE, returns a matrix with three columns, where the first two columns define the pair of genes/proteins and the third column defines the prediction/score. If FALSE, returns an adjacency matrix encoding the same information (but may be sparse depending on how many predictions were made in the original predict call.)

...

Additional parameters for consistency with generic.

Details

predict.ProtWeaver returns a ProtWeb object, which bundles some methods to make formatting and printing of results slightly nicer. This method extracts data form the ProtWeb object.

If AsDf=TRUE, the return data is a 3xN data.frame, with columns 3 showing the prediction for the pair of genes/proteins specified in columns 1 and 2. This format is when predictions are made on small number of pairs (meaning the resulting adjacency matrix is sparse).

If AsDf=FALSE, the return data is a NxN adjacency matrix, with entry i,j containing the prediction for genes i and j.

Value

Either a data.frame or a matrix.

Author

Aidan Lakshman ahl27@pitt.edu

See also

Examples

##############
## Prediction with built-in model and data 
###############

exData <- get(data("ExampleStreptomycesData"))

# Subset isn't necessary but is faster for a working example
pw <- ProtWeaver(exData$Genes[1:10])
#> Warning: Disabling MirrorTree-based algorithms. Input list must include inputs of type "dendrogram" for MT algorithms. Consult the documentation for more info.

protweb <- predict(pw, method='Jaccard')
#> Calculating P/A profiles:
#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#> Running ProfileDCA:
#>   Finding topology...
#>   || Done.
#>   Eliminating edges close to zero (may take a moment)...
#>   Refining edge weights...
#>   || Done.
#> Running Jaccard:
#>   |                                                                              |                                                                      |   0%  |                                                                              |===                                                                   |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  13%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  22%  |                                                                              |=================                                                     |  24%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  33%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  38%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  42%  |                                                                              |===============================                                       |  44%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  49%  |                                                                              |====================================                                  |  51%  |                                                                              |=====================================                                 |  53%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  58%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  69%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  73%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  78%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%
#> Running Hamming:
#>   |                                                                              |                                                                      |   0%  |                                                                              |===                                                                   |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  13%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  22%  |                                                                              |=================                                                     |  24%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  33%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  38%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  42%  |                                                                              |===============================                                       |  44%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  49%  |                                                                              |====================================                                  |  51%  |                                                                              |=====================================                                 |  53%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  58%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  69%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  73%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  78%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%
#> Running MutualInformation:
#>   |                                                                              |                                                                      |   0%  |                                                                              |===                                                                   |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  13%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  22%  |                                                                              |=================                                                     |  24%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  33%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  38%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  42%  |                                                                              |===============================                                       |  44%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  49%  |                                                                              |====================================                                  |  51%  |                                                                              |=====================================                                 |  53%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  58%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  69%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  73%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  78%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%
#> Running Coloc:
#>   |                                                                              |                                                                      |   0%  |                                                                              |==                                                                    |   2%  |                                                                              |===                                                                   |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   9%  |                                                                              |========                                                              |  11%  |                                                                              |=========                                                             |  13%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  18%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  22%  |                                                                              |=================                                                     |  24%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  29%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  33%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  38%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  42%  |                                                                              |===============================                                       |  44%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  49%  |                                                                              |====================================                                  |  51%  |                                                                              |=====================================                                 |  53%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  58%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  69%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  73%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  78%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  93%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  98%  |                                                                              |======================================================================| 100%
#> Calculating additional P/A Statistics...
#> Predicting with Ensemble method...
#> Done.
#> 
#> Time difference of 2.14 seconds.

# print out results as an adjacency matrix
GetProtWebData(protweb)
#>            1          2          3          4          5         6          7
#> 1  1.0000000 0.39977312 0.29051862 0.24212198 0.15859257 0.5806859 0.12798610
#> 2  0.3997731 1.00000000 0.13100765 0.15225417 0.10856788 0.2330527 0.10323494
#> 3  0.2905186 0.13100765 1.00000000 0.99982883 0.07591308 0.1601643 0.10597995
#> 4  0.2421220 0.15225417 0.99982883 1.00000000 0.06367025 0.1776079 0.07127068
#> 5  0.1585926 0.10856788 0.07591308 0.06367025 1.00000000 0.1361201 0.08176609
#> 6  0.5806859 0.23305267 0.16016427 0.17760794 0.13612013 1.0000000 0.12500786
#> 7  0.1279861 0.10323494 0.10597995 0.07127068 0.08176609 0.1250079 1.00000000
#> 8  0.1069398 0.12737124 0.10620673 0.10809579 0.08224316 0.1214757 0.99565868
#> 9  0.1695797 0.10552162 0.10804297 0.08227619 0.06591927 0.1316778 0.94517017
#> 10 0.1351764 0.08093309 0.08395500 0.10219318 0.08920908 0.0949550 0.76167646
#>             8          9         10
#> 1  0.10693984 0.16957966 0.13517643
#> 2  0.12737124 0.10552162 0.08093309
#> 3  0.10620673 0.10804297 0.08395500
#> 4  0.10809579 0.08227619 0.10219318
#> 5  0.08224316 0.06591927 0.08920908
#> 6  0.12147573 0.13167781 0.09495500
#> 7  0.99565868 0.94517017 0.76167646
#> 8  1.00000000 0.90979779 0.63626273
#> 9  0.90979779 1.00000000 0.99972807
#> 10 0.63626273 0.99972807 1.00000000

# print out results as a pairwise data.frame
GetProtWebData(protweb, AsDf=TRUE)
#>    Gene1 Gene2 Prediction
#> 1      1     2 0.39977312
#> 2      1     3 0.29051862
#> 3      2     3 0.13100765
#> 4      1     4 0.24212198
#> 5      2     4 0.15225417
#> 6      3     4 0.99982883
#> 7      1     5 0.15859257
#> 8      2     5 0.10856788
#> 9      3     5 0.07591308
#> 10     4     5 0.06367025
#> 11     1     6 0.58068589
#> 12     2     6 0.23305267
#> 13     3     6 0.16016427
#> 14     4     6 0.17760794
#> 15     5     6 0.13612013
#> 16     1     7 0.12798610
#> 17     2     7 0.10323494
#> 18     3     7 0.10597995
#> 19     4     7 0.07127068
#> 20     5     7 0.08176609
#> 21     6     7 0.12500786
#> 22     1     8 0.10693984
#> 23     2     8 0.12737124
#> 24     3     8 0.10620673
#> 25     4     8 0.10809579
#> 26     5     8 0.08224316
#> 27     6     8 0.12147573
#> 28     7     8 0.99565868
#> 29     1     9 0.16957966
#> 30     2     9 0.10552162
#> 31     3     9 0.10804297
#> 32     4     9 0.08227619
#> 33     5     9 0.06591927
#> 34     6     9 0.13167781
#> 35     7     9 0.94517017
#> 36     8     9 0.90979779
#> 37     1    10 0.13517643
#> 38     2    10 0.08093309
#> 39     3    10 0.08395500
#> 40     4    10 0.10219318
#> 41     5    10 0.08920908
#> 42     6    10 0.09495500
#> 43     7    10 0.76167646
#> 44     8    10 0.63626273
#> 45     9    10 0.99972807